DX_Project/LGDX프로젝트_모델_db/server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
google-generativeai
